requirements.txt
apache-airflow==2.5.1
google-cloud-storage
google-cloud-bigquery
faker
dbt-core==1.5.0
dbt-bigquery
pandas
fastapi
uvicorn

